+++ R/q_dtm.R
@@ -20,6 +20,8 @@
 #' @param keep.hyphen logical.  If \code{TRUE} hyphens are retained in the terms
 #' (e.g., "math-like" is kept as "math-like"), otherwise they become a split for
 #' terms (e.g., "math-like" is converted to "math" & "like").
+#' @param ngrams A vector of ngrams (multiple wrds with spaces).  Using this
+#' option results in the ngrams that will be retained in the matrix.
 #' @param \ldots Additional arguments passed to \code{\link[quanteda]{dfm}}.
 #' @return Returns a \code{\link[tm]{DocumentTermMatrix}}.
 #' @keywords dtm DocumentTermMatrix
@@ -33,9 +35,26 @@
 #'
 #' (x2 <- with(presidential_debates_2012, q_dtm_stem(dialogue, paste(time, tot, sep = "_"))))
 #' remove_stopwords(x2, stem=TRUE)
-q_dtm <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FALSE, ...){
+#'
+#' bigrams <- c('make sure', 'governor romney', 'mister president',
+#'     'united states', 'middle class', 'middle east', 'health care',
+#'     'american people', 'dodd frank', 'wall street', 'small business')
+#'
+#' grep(" ", x$dimnames$Terms, value = TRUE) #no ngrams
+#'
+#' (x3 <- with(presidential_debates_2012,
+#'     q_dtm(dialogue, paste(time, tot, sep = "_"), ngrams = bigrams)
+#' ))
+#'
+#' grep(" ", x3$dimnames$Terms, value = TRUE) #ngrams
+q_dtm <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FALSE,
+    ngrams = NULL, ...){
 
     if (!keep.hyphen) text <- gsub("-", " ", text)
+    if (!is.null(ngrams)) {
+        ngrams_replace <- gsub('\\s+', 'gofastrseparatorgofastr', ngrams)
+        text <- .mgsub(ngrams, ngrams_replace, text)
+    }
 
     if (length(unique(docs)) != length(text)){
         text <- unlist(lapply(split(text, docs), paste, collapse = " "))
@@ -44,6 +63,8 @@ q_dtm <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FALSE, 
 
     out <- quanteda::convert(quanteda::dfm(text, stem = FALSE, verbose = FALSE, removeNumbers = FALSE, ...), to = to)
     row.names(out) <- docs
+    if (!is.null(ngrams))colnames(out) <- gsub('gofastrseparatorgofastr', ' ', colnames(out))
+
     out
 }
 
@@ -65,9 +86,14 @@ q_dtm <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FALSE, 
 
 #' @export
 #' @rdname q_dtm
-q_dtm_stem <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FALSE, ...){
+q_dtm_stem <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FALSE,
+    ngrams = NULL, ...){
 
     if (!keep.hyphen) text <- gsub("-", " ", text)
+    if (!is.null(ngrams)) {
+        ngrams_replace <- gsub('\\s+', 'gofastrseparatorgofastr', ngrams)
+        text <- .mgsub(ngrams, ngrams_replace, text)
+    }
 
     if (length(unique(docs)) != length(text)){
         text <- unlist(lapply(split(text, docs), paste, collapse = " "))
@@ -76,6 +102,8 @@ q_dtm_stem <- function(text, docs = seq_along(text), to = "tm", keep.hyphen = FA
 
     out <- quanteda::convert(quanteda::dfm(text, stem = TRUE, verbose = FALSE, removeNumbers = FALSE, ...), to = to)
     row.names(out) <- docs
+    if (!is.null(ngrams))colnames(out) <- gsub('gofastrseparatorgofastr', ' ', colnames(out))
+
     out
 }